misc/sanity-check/challenge/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "brun0ne",email = "brunonblok@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "limits (>=5.1.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
